+++ src/main/java/com/ssssogong/issuemanager/domain/role/Role.java
@@ -6,15 +6,23 @@ import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.ToString;
+import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.authority.GrantedAuthoritiesContainer;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
 
+import java.util.ArrayList;
 import java.util.Collection;
+import java.util.List;
 
 @Entity
 @AllArgsConstructor
 @NoArgsConstructor
 @ToString
-public abstract class Role implements GrantedAuthoritiesContainer {
+public class Role implements GrantedAuthoritiesContainer {
+    public Role(Collection<Privilege> privileges){
+        allowedPrivileges = privileges;
+    }
+
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Getter
@@ -26,4 +34,20 @@ public abstract class Role implements GrantedAuthoritiesContainer {
     @Column
     @Getter
     protected Collection<Privilege> allowedPrivileges;
+
+    @Override
+    public Collection<GrantedAuthority> getGrantedAuthorities() {
+        // 자신의 역할 + Privilege를 담아서 반환한다
+        String roleName = this.getClass().getName();
+        List<GrantedAuthority> authorities = new ArrayList<>();
+        if(!roleName.equals("Role"))
+            authorities.add(new SimpleGrantedAuthority("ROLE_" + roleName));
+        authorities.addAll(
+                allowedPrivileges
+                        .stream()
+                        .map(privilege -> new SimpleGrantedAuthority(privilege.name()))
+                        .toList()
+        );
+        return authorities;
+    }
 }
